When creating an availability, add the availability object to session only after all fields are filled in
This is similar to an issue encountered previously --- when SQLAlchemy issues an INSERT statement from session.add(), any updates to the object before session.commit() do not get reflected in an updated INSERT statement.

This means that if you session.add() an empty object (which was done previously), and a column has a NOT NULL constraint (which is added by SQLAlchemy by default), it will reflect as inserting a row with empty values. This is obviously disallowed, and this error shows up in a failure of one (or some) of the columns' NOT NULL constraint.